@@ -3,41 +3,43 @@
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
-  "homepage": "http://rajithaBK.github.io/embeddable-meetings-widget",
+  "homepage": "http://wxsd-sales.github.io/embeddable-meetings-widget",
   "dependencies": {
-    "@babel/core": "^7.4.5",
+    "@babel/core": "^7.25.2",
     "@babel/polyfill": "^7.12.1",
-    "@nrwl/react": "^13.8.4",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "@webex/internal-plugin-locus": "^1.156.0",
-    "@webex/widgets": "^1.20.6",
-    "contentful": "^9.1.12",
-    "core-js": "^3.21.1",
+    "@nrwl/react": "^19.8.1",
+    "@testing-library/jest-dom": "^6.5.0",
+    "@testing-library/react": "^16.0.1",
+    "@testing-library/user-event": "^14.5.2",
+    "@webex/internal-plugin-locus": "^3.4.0",
+    "@webex/widgets": "^1.27.0",
+    "contentful": "^11.0.3",
+    "core-js": "^3.38.1",
     "crypto-browserify": "^3.12.0",
-    "css-loader": "^5.2.7",
+    "css-loader": "^7.1.2",
     "file-loader": "^6.2.0",
     "fs": "0.0.1-security",
-    "gh-pages": "^3.2.3",
+    "gh-pages": "^6.1.1",
     "http-browserify": "^1.7.0",
     "https-browserify": "^1.0.0",
     "os-browserify": "^0.3.0",
     "path": "^0.12.7",
     "pnp-webpack-plugin": "^1.7.0",
-    "qs": "^6.10.1",
-    "query-string": "^7.0.1",
+    "qs": "^6.13.0",
+    "query-string": "^9.1.0",
     "raw-loader": "^4.0.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-draggable": "^4.4.4",
-    "react-scripts": "^5.0.0",
-    "stream": "0.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-draggable": "^4.4.6",
+    "react-scripts": "^5.0.1",
+    "regenerator-runtime": "^0.14.1",
+    "stream": "0.0.3",
     "stream-browserify": "^3.0.0",
-    "style-loader": "^2.0.0",
+    "style-loader": "^4.0.0",
+    "url": "^0.11.4",
     "url-loader": "^4.1.1",
-    "web-vitals": "^1.0.1",
-    "webex": "^1.150.0"
+    "web-vitals": "^4.2.3",
+    "webex": "^3.4.0"
   },
   "resolutions": {
     "react-error-overlay": "6.0.9"
@@ -53,25 +55,23 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@types/react": "^16.9.35",
-    "@types/react-dom": "^16.9.8",
-    "@types/styled-components": "^5.1.0",
-    "awesome-typescript-loader": "^5.2.1",
+    "@types/react": "^18.3.9",
+    "@types/react-dom": "^18.3.0",
+    "@types/styled-components": "^5.1.34",
     "babel-polyfill": "^6.26.0",
-    "clean-webpack-plugin": "^3.0.0",
-    "extract-text-webpack-plugin": "^3.0.2",
-    "live-server": "^1.2.1",
-    "mini-css-extract-plugin": "^0.9.0",
+    "clean-webpack-plugin": "^4.0.0",
+    "live-server": "^1.2.2",
+    "mini-css-extract-plugin": "^2.9.1",
     "process": "^0.11.10",
-    "react-dom": "^17.0.2",
-    "react-error-overlay": "^6.0.9",
+    "react-dom": "^18.3.1",
+    "react-error-overlay": "^6.0.11",
     "speed-measure-webpack-plugin": "^1.5.0",
-    "ts-loader": "^7.0.3",
-    "typescript": "^3.8.3",
-    "webpack": "^5.70.0",
-    "webpack-cli": "^5.0.1",
-    "webpack-dev-server": "^3.11.0",
-    "webpack-merge": "^4.2.2"
+    "ts-loader": "^9.5.1",
+    "typescript": "^4.9.5",
+    "webpack": "^5.95.0",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^5.1.0",
+    "webpack-merge": "^6.0.1"
   },
   "eslintConfig": {
     "extends": [
